Move `QuestionActions` component
