+++ frontend/src/metabase/query_builder/components/view/ViewHeader/ViewHeader.jsx
@@ -17,13 +17,13 @@ import SavedQuestionHeaderButton from "metabase/query_builder/components/SavedQu
 import { navigateBackToDashboard } from "metabase/query_builder/actions";
 import { MODAL_TYPES } from "metabase/query_builder/constants";
 import { getDashboard } from "metabase/query_builder/selectors";
-import QuestionActions from "../../QuestionActions";
 import { FilterHeaderToggle, FilterHeader } from "../QuestionFilters";
 import { QuestionSummarizeWidget } from "../QuestionSummaries";
 import { ConvertQueryButton } from "./components/ConvertQueryButton";
 import { HeadBreadcrumbs } from "./components/HeaderBreadcrumbs";
 import { FilterHeaderButton } from "./components/FilterHeaderButton";
 import { ExploreResultsLink } from "./components/ExploreResultsLink";
+import { QuestionActions } from "./components/QuestionActions";
 import { QuestionNotebookButton } from "./components/QuestionNotebookButton";
 import { QuestionDataSource } from "./components/QuestionDataSource";
 import { QuestionDescription } from "./components/QuestionDescription";
